Updated player logic
The game now registers an error on the player's part if their answer does
not match "rock" "paper" or "scissors" (player answer is not case sensitive).
This now correctly does not register a win or a loss for either the
player or computer, but does increase the games played counter.